Seperate settings tabs (ProfileTab and CategoriesTab) out of Settings component.
