+++ src/pages/Settings/Settings.jsx
@@ -1,14 +1,5 @@
-import {
-  Pencil,
-  ArrowBigLeft,
-  Check,
-  X,
-  GripVertical,
-  Eye,
-  EyeOff,
-} from "lucide-react";
-import { DragDropContext, Droppable, Draggable } from "@hello-pangea/dnd";
-import { useState, useEffect, useRef, useCallback } from "react";
+import { ArrowBigLeft } from "lucide-react";
+import { useState, useEffect, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import {
@@ -19,13 +10,11 @@ import {
   checkUsernameExists,
   deleteUserAccount,
 } from "../../services/userService";
-import {
-  saveUserCategoryPreferences,
-  getAllCategoriesForManagement,
-} from "../../services/categoryPreferencesService";
 import { useUnsavedChanges } from "../../hooks/ui/useUnsavedChanges";
 import LoadingAcorn from "../../components/LoadingAcorn/LoadingAcorn";
 import ConfirmationModal from "../../components/ConfirmationModal/ConfirmationModal";
+import ProfileTab from "./ProfileTab";
+import CategoriesTab from "./CategoriesTab";
 import "./Settings.css";
 
 const Settings = () => {
@@ -381,7 +370,6 @@ const Settings = () => {
                 saveMessage={saveMessage}
                 setSaveMessage={setSaveMessage}
                 onUnsavedChangesChange={setCategoriesHasUnsavedChanges}
-                navigateWithConfirmation={navigateWithConfirmation}
               />
             )}
           </div>
@@ -409,492 +397,6 @@ const Settings = () => {
           />
         </>
       )}
-    </div>
-  );
-};
-
-const ProfileTab = ({
-  profileData,
-  isEditingUsername,
-  tempUsername,
-  usernameError,
-  isEditingFirstName,
-  tempFirstName,
-  isEditingLanguage,
-  tempLanguage,
-  usernameInputRef,
-  firstNameInputRef,
-  usernameContainerRef,
-  firstNameContainerRef,
-  languageContainerRef,
-  handleEditUsername,
-  handleSaveUsername,
-  handleCancelUsername,
-  handleEditFirstName,
-  handleSaveFirstName,
-  handleCancelFirstName,
-  handleChangePassword,
-  handleEditLanguage,
-  handleSaveLanguage,
-  handleCancelLanguage,
-  handleDeleteAccount,
-  setTempUsername,
-  setUsernameError,
-  setTempFirstName,
-  setTempLanguage,
-  t,
-}) => {
-  return (
-    <div className="profile-settings-container">
-      <div className="profile-settings">
-        <div className="profile-settings-column">
-          <div className="input-validation-wrapper" ref={firstNameContainerRef}>
-            <div className="floating-label-input">
-              <div className="relative-center">
-                <input
-                  ref={firstNameInputRef}
-                  id="first_name"
-                  type="text"
-                  value={
-                    isEditingFirstName ? tempFirstName : profileData.first_name
-                  }
-                  onChange={
-                    isEditingFirstName
-                      ? (e) => setTempFirstName(e.target.value)
-                      : undefined
-                  }
-                  className={`input   ${
-                    isEditingFirstName ? "input--edit" : ""
-                  }`}
-                  placeholder=" "
-                  readOnly={!isEditingFirstName}
-                />
-                {!isEditingFirstName ? (
-                  <button
-                    className="btn-unstyled btn-icon btn-icon-right"
-                    onClick={handleEditFirstName}
-                    aria-label={t("edit_first_name")}
-                  >
-                    <Pencil size={20} />
-                  </button>
-                ) : (
-                  <>
-                    <button
-                      className="btn-unstyled btn-icon btn-icon-green profile-settings-check"
-                      onClick={handleSaveFirstName}
-                      aria-label={t("save_changes")}
-                    >
-                      <Check size={20} />
-                    </button>
-                    <button
-                      className="btn-unstyled btn-icon btn-icon-red profile-settings-cancel"
-                      onClick={handleCancelFirstName}
-                      aria-label={t("cancel")}
-                    >
-                      <X size={20} />
-                    </button>
-                  </>
-                )}
-              </div>
-              <label htmlFor="first_name">{t("first_name")}</label>
-            </div>
-          </div>
-
-          <div className="input-validation-wrapper">
-            <div className="floating-label-input">
-              <div className="relative-center">
-                <input
-                  id="email"
-                  type="email"
-                  value={profileData.email}
-                  className="input  "
-                  placeholder=" "
-                  readOnly={true}
-                />
-              </div>
-              <label htmlFor="email">{t("email")}</label>
-            </div>
-          </div>
-        </div>
-
-        <div className="profile-settings-column">
-          <div
-            className="input-validation-wrapper"
-            style={{ position: "relative" }}
-            ref={usernameContainerRef}
-          >
-            <div className="floating-label-input">
-              <div className="relative-center">
-                <input
-                  ref={usernameInputRef}
-                  id="username"
-                  type="text"
-                  value={
-                    isEditingUsername ? tempUsername : profileData.username
-                  }
-                  onChange={
-                    isEditingUsername
-                      ? (e) => {
-                          setTempUsername(e.target.value);
-                          setUsernameError("");
-                        }
-                      : undefined
-                  }
-                  className={`input   ${
-                    isEditingUsername ? "input--edit" : ""
-                  } ${usernameError ? "input--error" : ""}`}
-                  placeholder=" "
-                  readOnly={!isEditingUsername}
-                />
-                {!isEditingUsername ? (
-                  <button
-                    className="btn-unstyled btn-icon btn-icon-right"
-                    onClick={handleEditUsername}
-                    aria-label={t("edit_username")}
-                  >
-                    <Pencil size={20} />
-                  </button>
-                ) : (
-                  <>
-                    <button
-                      className="btn-unstyled btn-icon btn-icon-green profile-settings-check"
-                      onClick={handleSaveUsername}
-                      aria-label={t("save_changes")}
-                    >
-                      <Check size={20} />
-                    </button>
-                    <button
-                      className="btn-unstyled btn-icon btn-icon-red profile-settings-cancel"
-                      onClick={handleCancelUsername}
-                      aria-label={t("cancel")}
-                    >
-                      <X size={20} />
-                    </button>
-                  </>
-                )}
-              </div>
-              <label htmlFor="username">{t("username")}</label>
-            </div>
-            {usernameError && (
-              <span className="error-message-small error-message-absolute">
-                {usernameError}
-              </span>
-            )}
-          </div>
-
-          <div className="input-validation-wrapper">
-            <div className="floating-label-input">
-              <div className="relative-center">
-                <input
-                  id="password"
-                  type="password"
-                  value="**************"
-                  className="input"
-                  placeholder=" "
-                  readOnly={true}
-                />
-                <button
-                  className="btn btn-icon btn-icon-right"
-                  onClick={handleChangePassword}
-                  aria-label={t("change_password")}
-                >
-                  <Pencil size={20} />
-                </button>
-              </div>
-              <label htmlFor="password">{t("password")}</label>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div className="profile-settings-language flex-column-center">
-        <h2 className="forta-small">{t("preferred_language").toUpperCase()}</h2>
-
-        <div
-          className="profile-settings-language-container"
-          ref={languageContainerRef}
-        >
-          {!isEditingLanguage ? (
-            <div className="language-wrapper">
-              <span
-                className={`language${
-                  (profileData.preferred_language || "en") === "en"
-                    ? " selected"
-                    : ""
-                }`}
-              >
-                EN
-              </span>
-              |
-              <span
-                className={`language${
-                  (profileData.preferred_language || "en") === "de"
-                    ? " selected"
-                    : ""
-                }`}
-              >
-                DE
-              </span>
-            </div>
-          ) : (
-            <div className="language-wrapper">
-              <button
-                className={`btn-unstyled language${
-                  tempLanguage === "en" ? " selected" : ""
-                }`}
-                onClick={() => setTempLanguage("en")}
-              >
-                EN
-              </button>
-              |
-              <button
-                className={`btn-unstyled language${
-                  tempLanguage === "de" ? " selected" : ""
-                }`}
-                onClick={() => setTempLanguage("de")}
-              >
-                DE
-              </button>
-            </div>
-          )}
-
-          {!isEditingLanguage ? (
-            <button
-              className="btn btn-unstyled profile-settings-check"
-              onClick={handleEditLanguage}
-              aria-label={t("edit_language")}
-            >
-              <Pencil size={20} />
-            </button>
-          ) : (
-            <>
-              <button
-                className="btn btn-icon btn-icon-green profile-settings-check"
-                onClick={handleSaveLanguage}
-                aria-label={t("save_changes", "Save changes")}
-              >
-                <Check size={20} />
-              </button>
-              <button
-                className="btn btn-icon btn-icon-red profile-settings-cancel"
-                onClick={handleCancelLanguage}
-                aria-label={t("cancel", "Cancel")}
-              >
-                <X size={20} />
-              </button>
-            </>
-          )}
-        </div>
-      </div>
-
-      <button
-        onClick={handleDeleteAccount}
-        className="btn btn-action btn-danger"
-      >
-        {t("delete_account")}
-      </button>
-    </div>
-  );
-};
-
-const CategoriesTab = ({
-  t,
-  saveMessage,
-  setSaveMessage,
-  onUnsavedChangesChange,
-}) => {
-  const [categories, setCategories] = useState([]);
-  const [categoriesLoading, setCategoriesLoading] = useState(true);
-  const [categoryPreferences, setCategoryPreferences] = useState([]);
-  const [originalCategoryPreferences, setOriginalCategoryPreferences] =
-    useState([]);
-  const [preferencesLoading, setPreferencesLoading] = useState(false);
-  const { i18n } = useTranslation();
-
-  // Load all categories for management (including hidden ones)
-  useEffect(() => {
-    const loadAllCategories = async () => {
-      try {
-        setCategoriesLoading(true);
-        const allCategories = await getAllCategoriesForManagement(
-          i18n.language
-        );
-        setCategories(allCategories);
-      } catch (error) {
-        console.error("Error loading categories for management:", error);
-      } finally {
-        setCategoriesLoading(false);
-      }
-    };
-
-    loadAllCategories();
-  }, [i18n.language]);
-
-  useEffect(() => {
-    // Categories already come with preference data from getAllCategoriesForManagement
-    if (categories.length > 0) {
-      setCategoryPreferences([...categories]);
-      setOriginalCategoryPreferences([...categories]);
-    }
-  }, [categories]);
-
-  // Check if preferences have changed
-  const hasUnsavedChanges = useCallback(() => {
-    if (categoryPreferences.length !== originalCategoryPreferences.length) {
-      return true;
-    }
-
-    return categoryPreferences.some((pref, index) => {
-      const original = originalCategoryPreferences[index];
-      return (
-        pref.isVisible !== original.isVisible || pref.order !== original.order
-      );
-    });
-  }, [categoryPreferences, originalCategoryPreferences]);
-
-  // Notify parent component about unsaved changes
-  useEffect(() => {
-    onUnsavedChangesChange?.(hasUnsavedChanges());
-  }, [hasUnsavedChanges, onUnsavedChangesChange]);
-
-  const toggleVisibility = (categoryId) => {
-    setCategoryPreferences((prev) =>
-      prev.map((cat) =>
-        cat.id === categoryId ? { ...cat, isVisible: !cat.isVisible } : cat
-      )
-    );
-  };
-
-  const handleDragEnd = (result) => {
-    if (!result.destination) return;
-
-    const items = Array.from(categoryPreferences);
-    const [reorderedItem] = items.splice(result.source.index, 1);
-    items.splice(result.destination.index, 0, reorderedItem);
-
-    // Update order values
-    items.forEach((cat, index) => {
-      cat.order = index;
-    });
-
-    setCategoryPreferences(items);
-  };
-
-  const handleSavePreferences = async () => {
-    try {
-      setPreferencesLoading(true);
-      setSaveMessage("");
-
-      await saveUserCategoryPreferences(categoryPreferences);
-
-      // Update original preferences to match current state
-      setOriginalCategoryPreferences([...categoryPreferences]);
-
-      setSaveMessage(t("category_preferences_saved"));
-      setTimeout(() => setSaveMessage(""), 3000);
-    } catch (error) {
-      console.error("Error saving category preferences:", error);
-      setSaveMessage(t("category_preferences_error"));
-      setTimeout(() => setSaveMessage(""), 3000);
-    } finally {
-      setPreferencesLoading(false);
-    }
-  };
-
-  if (categoriesLoading || categoryPreferences.length === 0) {
-    return <LoadingAcorn />;
-  }
-
-  return (
-    <div className="flex-column">
-      <div className="flex-column-center">
-        <p className="grey-small">{t("category_management_description")}</p>
-      </div>
-
-      <DragDropContext onDragEnd={handleDragEnd}>
-        <Droppable droppableId="categories">
-          {(provided, snapshot) => (
-            <div
-              className={`category-list ${
-                snapshot.isDraggingOver ? "drag-over" : ""
-              }`}
-              {...provided.droppableProps}
-              ref={provided.innerRef}
-            >
-              {categoryPreferences.map((category, index) => (
-                <Draggable
-                  key={category.id}
-                  draggableId={category.id.toString()}
-                  index={index}
-                >
-                  {(provided, snapshot) => (
-                    <div
-                      ref={provided.innerRef}
-                      {...provided.draggableProps}
-                      className={`category-item ${
-                        !category.isVisible ? "category-hidden" : ""
-                      } ${snapshot.isDragging ? "dragging" : ""}`}
-                    >
-                      <div
-                        {...provided.dragHandleProps}
-                        className="drag-handle"
-                      >
-                        <GripVertical size={16} />
-                      </div>
-
-                      <div className="category-info">
-                        <span className="category-name">{category.label}</span>
-                        {/* {category.isSystem && (
-                          <span className="category-system-badge">
-                            {t("system")}
-                          </span>
-                        )} */}
-                      </div>
-
-                      <button
-                        className="btn-unstyled category-visibility-toggle"
-                        onClick={() => toggleVisibility(category.id)}
-                        aria-label={
-                          category.isVisible
-                            ? t("hide_category")
-                            : t("show_category")
-                        }
-                      >
-                        {category.isVisible ? (
-                          <Eye size={16} />
-                        ) : (
-                          <EyeOff size={16} />
-                        )}
-                      </button>
-                    </div>
-                  )}
-                </Draggable>
-              ))}
-              {provided.placeholder}
-            </div>
-          )}
-        </Droppable>
-      </DragDropContext>
-
-      <div className="success-message-wrapper">
-        <span
-          className={`red-small ${
-            saveMessage?.includes("Error") ? "error-message" : ""
-          }`}
-        >
-          {saveMessage || "\u00A0"}
-        </span>
-      </div>
-
-      <div className="action-buttons">
-        <button
-          className="btn btn-action btn-primary"
-          onClick={handleSavePreferences}
-          disabled={preferencesLoading}
-        >
-          {preferencesLoading ? t("saving") : t("save_category_preferences")}
-        </button>
-      </div>
     </div>
   );
 };
